+++ CHANGELOG.md
@@ -12,8 +12,9 @@
 
 ### Changed
 
-- Remove the websocket message configuration option
+- Remove the WebSocket message configuration option
 - Use the return value of `HuangYi\Shadowfax\Contracts\WebSocket\Handler::messageClass()` method first to create the message instance (if any)
+- Close the WebSocket connection when receiving an invalid message
 
 
 ## v2.2.0